add and delete

+++ MedicalClinic/src/main/java/com/example/demo/TreatmentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 
 //CRUD refers Create, Read, Update, Delete
 
-public interface TreatmentRepository extends CrudRepository<Treatment, Integer> {
+public interface TreatmentRepository extends CrudRepository<Treatment, Long> {
 	List<Treatment> findByAppointmentId(Long appointment_id);
 }
